Move StableConfigMappingExceptionTest
